refactor: move Contract pages under Projects
