+++ src/pages/Projects/EditableContractComponents/InspectionForm.tsx
